Fix pp3 deletion when clearing processing params
Fix comparison of original vs. current rank. The rank value can be -1,
which should be interpreted as 0. This allows proper assessment if the
rank has been changed. The pp3 is deleted if and only if the rank, color
label, and trash status are unchanged. pp3s were not getting deleted
because of the mismatch between the original rank (-1) and current rank
(0).

Also consider the rank and color label in the metadata and XMP sidecars.
If the rank and color label have not been edited in the current session
and the current values match the ones in the metadata and XMP sidecars,
the pp3 can be deleted.