Fix Nikon D5 black level when using LibRaw
With LibRaw, the black level is provided in the global black variable
instead of the per-channel array. Black levels in camconst.json override
the global black level instead of adding to the black level as with the
per-channel levels. LibRaw and dcraw have the same black level for the
D5, so it is sufficient to remove the black level from camconst.json.