Fix blown images when using LibRaw
dcraw code only normalizes the pre-multipliers for the number of colors
in the raw image. LibRaw can set the pre-multipliers for a greater
number of colors. For example, it can set the 4th multiplier for a
3-color raw. This causes the dcraw code to produce incorrect
multipliers.

Remove the 4th multiplier if the number of colors is less than 4 and the
multiplier is equal to the 2nd multiplier. This lets the multiplier be
set equal to the 2nd one after normalization. For images where the
multipliers are not equal, print a warning.